fix: change exit code type from float32 to int
- Updated ExitCode field in SandboxProcessResponseDto from float32 to int
- Maintained JSON tag consistency with other fields
- Preserved base64 encoding documentation for stdout and stderr fields

+++ model_sandbox_process_response_dto.go
@@ -8,6 +8,7 @@
  */
 
 package lybic
+
 // SandboxProcessResponseDto struct for SandboxProcessResponseDto
 type SandboxProcessResponseDto struct {
 	// stdout as base64-encoded bytes
@@ -15,5 +16,5 @@ type SandboxProcessResponseDto struct {
 	// stderr as base64-encoded bytes
 	StderrBase64 string `json:"stderrBase64,omitempty"`
 	// Exit code
-	ExitCode float32 `json:"exitCode"`
+	ExitCode int `json:"exitCode"`
 }
